moved the hooks into a hooks directory so only db calls are in api directory
