+++ src/hooks/useGetThemes.js
@@ -1,0 +1,22 @@
+import {useState, useEffect } from "react";
+import {getThemes} from '../api/index.js';
+
+export const useGetThemes = () => {
+      const [themes, setThemes] = useState([]);
+
+      useEffect(() => {
+        let userId = 1234567;
+
+        async function fetchData() {
+          const result = await getThemes(userId);
+          setThemes(result);
+        }
+
+        fetchData();
+        return () => {}
+      }, [themes]);
+
+      return (
+        [themes]
+      );
+  }
